requirements.txt
feedparser==5.2.1
peewee==2.8.0
future==0.15.2
python_telegram_bot
simplejson==3.8.2
